Basic translation from high level notes to low levels implemented.
Fixed mistake in tempoToDTime, as the tempo gives the number of whole
notes in a minute.  # Please enter the commit message for your
changes. Lines starting # with '#' will be ignored, and an empty
message aborts the commit.  # On branch master # Your branch is
up-to-date with 'origin/master'.  # # Changes to be committed: #
modified: RCMA/Global/Clock.hs # modified: RCMA/Translator/Note.hs #